Refactor to use react-portal
Moving to react-portal gives us the advantage that context is properly preserved through the React tree. react-portal is a small library that just moves a subtree into its own DOM element at the end of body, and react-tether looks like a great use case for it while solving a problem with react-tether at the same time.